+++ src/googleapis_without_429/__init__.py
@@ -3,7 +3,7 @@
 from importlib.metadata import PackageNotFoundError, version
 
 from googleapis_without_429.core import WeightedSlidingWindow
-from googleapis_without_429.errors import is_rate_limited
+from googleapis_without_429.errors import QuotaTimeoutError, is_rate_limited
 from googleapis_without_429.limiter import QuotaLimiter
 from googleapis_without_429.profiles import DRIVE, SHEETS, ApiProfile
 from googleapis_without_429.session import RateLimitedSession
@@ -18,6 +18,7 @@ __all__ = [
     "SHEETS",
     "ApiProfile",
     "QuotaLimiter",
+    "QuotaTimeoutError",
     "RateLimitedSession",
     "WeightedSlidingWindow",
     "__version__",